refactor(notifications): split NotificationPreferencesForm under the line cap (#544)
Extract the collapsible per-type advanced settings section (query,
availableTypes derivation, five per-type helpers, and its template) into
a co-located NotificationTypeSettings.svelte. The per-type settings map
stays two-way via $bindable; global channels / disabled / telegram flags
pass down read-only. Behavior-preserving extraction only.

910 -> 567 lines (child: 374).
